test(cypress): trim reporting and test management runtime
